+++ src/type/type.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Injectable,
   InternalServerErrorException,
   NotFoundException,
@@ -6,10 +7,14 @@ import {
 import { TypeRepository } from './type.repository';
 import { AddTypeDTO } from './dto/add-type.dto';
 import { EditTypeDTO } from './dto/edit-type.dto';
+import { ProjectService } from '../project/project.service';
 
 @Injectable()
 export class TypeService {
-  constructor(private readonly typeRepo: TypeRepository) {}
+  constructor(
+    private readonly typeRepo: TypeRepository,
+    private readonly projectService: ProjectService,
+  ) {}
 
   getAll() {
     return this.typeRepo.getAll();
@@ -27,6 +32,26 @@ export class TypeService {
     return response;
   }
 
+  async checkType(id: number) {
+    const type = await this.typeRepo.getOneByIdOrFail(id);
+    if (!type) {
+      return null;
+    }
+    return type;
+  }
+
+  async validation(id: number, idProject: number) {
+    const checkType = await this.checkType(id);
+    if (!checkType) {
+      throw new NotFoundException();
+    }
+    const project = await this.projectService.getOneById(idProject);
+    if (!project) {
+      throw new NotFoundException('Project not Exist');
+    }
+    return true;
+  }
+
   async add(dto: AddTypeDTO) {
     try {
       const type = this.typeRepo.create(dto);
@@ -35,7 +60,18 @@ export class TypeService {
       throw new InternalServerErrorException();
     }
   }
-  async edit(id: number, dto: EditTypeDTO) {
+  async edit(id: number, idProject: number, dto: EditTypeDTO) {
+    const validation = this.validation(id, idProject);
+    if (!validation) {
+      return validation;
+    }
+    const statusExistProject = await this.typeRepo.countTypeInProject(
+      id,
+      idProject,
+    );
+    if (!statusExistProject) {
+      throw new BadRequestException('Status not Exist In Project');
+    }
     try {
       return await this.typeRepo.update(id, dto);
     } catch (e) {
@@ -43,10 +79,33 @@ export class TypeService {
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, idProject: number) {
+    const validation = this.validation(id, idProject);
+    if (!validation) {
+      return validation;
+    }
     try {
       await this.typeRepo.delete(id);
       return id;
+    } catch (e) {
+      throw new InternalServerErrorException();
+    }
+  }
+
+  async addTypeInProject(id: number, idProject: number) {
+    const validation = this.validation(id, idProject);
+    if (!validation) {
+      return validation;
+    }
+    const typeExistProject = await this.typeRepo.countTypeInProject(
+      id,
+      idProject,
+    );
+    if (typeExistProject) {
+      throw new BadRequestException('Type Exist In Project');
+    }
+    try {
+      return await this.typeRepo.update(id, { projectID: idProject });
     } catch (e) {
       throw new InternalServerErrorException();
     }
